get an example working with a label for the file name that called the New-PromethesMetric* functions.
Tried to use the DefaultRegistry.SetStaticLabels so these would apply to all metrics, but ran into type errors. This is working, it creates a new series (I think is the right term) with a New-GUID as the JobName. Then I set the static label to the filename, and the Instance is $env:ComputerName. You can use additional labels once you call New-PrometheusMetric* if needed